Add tests and updates for part 2 assignment
