+++ check-the-chess-EladBibi-main/.gitignore
@@ -1,0 +1,7 @@
+#
+# Untrack files
+#
+
+venv
+__pycache__
+.idea
