+++ check-the-chess-EladBibi-main/.replit
@@ -1,0 +1,2 @@
+language = "python3"
+run = "python3 -W ignore chess_gui.py"
