+++ check-the-chess-EladBibi-main/chess_gui.py
@@ -1,0 +1,213 @@
+#
+# The GUI engine for Python Chess
+#
+# Author: Boo Sung Kim, Eddie Sharick
+# Note: The pygame tutorial by Eddie Sharick was used for the GUI engine. The GUI code was altered by Boo Sung Kim to
+# fit in with the rest of the project.
+#
+
+import chess_engine
+import pygame as py
+import ai_engine
+from enums import Player
+import logging
+
+logging.basicConfig(
+    filename='chess_log.log',
+    level=logging.DEBUG,  # Changed to DEBUG for detailed logging
+    format='%(asctime)s - %(levelname)s - %(message)s'
+)
+
+WIDTH = HEIGHT = 512
+DIMENSION = 8
+SQ_SIZE = HEIGHT // DIMENSION
+MAX_FPS = 15
+IMAGES = {}
+colors = [py.Color("white"), py.Color("gray")]
+
+
+def load_images():
+    '''
+    Load images for the chess pieces
+    '''
+    for p in Player.PIECES:
+        IMAGES[p] = py.transform.scale(py.image.load("images/" + p + ".png"), (SQ_SIZE, SQ_SIZE))
+
+
+def draw_game_state(screen, game_state, valid_moves, square_selected):
+    ''' Draw the complete chess board with pieces
+
+    Keyword arguments:
+        :param screen       -- the pygame screen
+        :param game_state   -- the state of the current chess game
+    '''
+    draw_squares(screen)
+    highlight_square(screen, game_state, valid_moves, square_selected)
+    draw_pieces(screen, game_state)
+
+
+def draw_squares(screen):
+    ''' Draw the chess board with the alternating two colors
+
+    :param screen:          -- the pygame screen
+    '''
+    for r in range(DIMENSION):
+        for c in range(DIMENSION):
+            color = colors[(r + c) % 2]
+            py.draw.rect(screen, color, py.Rect(c * SQ_SIZE, r * SQ_SIZE, SQ_SIZE, SQ_SIZE))
+
+
+def draw_pieces(screen, game_state):
+    ''' Draw the chess pieces onto the board
+
+    :param screen:          -- the pygame screen
+    :param game_state:      -- the current state of the chess game
+    '''
+    for r in range(DIMENSION):
+        for c in range(DIMENSION):
+            piece = game_state.get_piece(r, c)
+            if piece is not None and piece != Player.EMPTY:
+                screen.blit(IMAGES[piece.get_player() + "_" + piece.get_name()],
+                            py.Rect(c * SQ_SIZE, r * SQ_SIZE, SQ_SIZE, SQ_SIZE))
+
+
+def highlight_square(screen, game_state, valid_moves, square_selected):
+    if square_selected != () and game_state.is_valid_piece(square_selected[0], square_selected[1]):
+        row = square_selected[0]
+        col = square_selected[1]
+        if (game_state.whose_turn() and game_state.get_piece(row, col).is_player(Player.PLAYER_1)) or \
+                (not game_state.whose_turn() and game_state.get_piece(row, col).is_player(Player.PLAYER_2)):
+            s = py.Surface((SQ_SIZE, SQ_SIZE))
+            s.set_alpha(100)
+            s.fill(py.Color("blue"))
+            screen.blit(s, (col * SQ_SIZE, row * SQ_SIZE))
+            s.fill(py.Color("green"))
+            for move in valid_moves:
+                screen.blit(s, (move[1] * SQ_SIZE, move[0] * SQ_SIZE))
+
+
+def draw_text(screen, text):
+    font = py.font.SysFont("Helvetica", 32, True, False)
+    text_object = font.render(text, False, py.Color("Black"))
+    text_location = py.Rect(0, 0, WIDTH, HEIGHT).move(WIDTH / 2 - text_object.get_width() / 2,
+                                                      HEIGHT / 2 - text_object.get_height() / 2)
+    screen.blit(text_object, text_location)
+
+
+def main():
+    knight_moves_count = 0
+    check_count = 0
+    human_player = ""
+
+    while True:
+        try:
+            number_of_players = input("How many players (1 or 2)?\n")
+            if int(number_of_players) == 1:
+                number_of_players = 1
+                while True:
+                    human_player = input("What color do you want to play (w or b)?\n")
+                    if human_player == "w" or human_player == "b":
+                        break
+                    else:
+                        print("Enter w or b.\n")
+                break
+            elif int(number_of_players) == 2:
+                number_of_players = 2
+                break
+            else:
+                print("Enter 1 or 2.\n")
+        except ValueError:
+            print("Enter 1 or 2.")
+
+    py.init()
+    screen = py.display.set_mode((WIDTH, HEIGHT))
+    clock = py.time.Clock()
+    game_state = chess_engine.game_state()
+    load_images()
+    running = True
+    square_selected = ()
+    player_clicks = []
+    valid_moves = []
+    game_over = False
+    ai = ai_engine.chess_ai()
+
+    # If AI plays first
+    if human_player == 'b':
+        ai_move = ai.minimax_black(game_state, 3, -100000, 100000, True, Player.PLAYER_1)
+        game_state.move_piece(ai_move[0], ai_move[1], True)
+
+    while running:
+        for e in py.event.get():
+            if e.type == py.QUIT:
+                running = False
+            elif e.type == py.MOUSEBUTTONDOWN:
+                if not game_over:
+                    location = py.mouse.get_pos()
+                    col = location[0] // SQ_SIZE
+                    row = location[1] // SQ_SIZE
+                    if square_selected == (row, col):
+                        square_selected = ()
+                        player_clicks = []
+                    else:
+                        square_selected = (row, col)
+                        player_clicks.append(square_selected)
+                    if len(player_clicks) == 2:
+                        if (player_clicks[1][0], player_clicks[1][1]) not in valid_moves:
+                            square_selected = ()
+                            player_clicks = []
+                            valid_moves = []
+                        else:
+                            moving_piece = game_state.get_piece(player_clicks[0][0], player_clicks[0][1])
+                            if moving_piece.get_name() == 'n':
+                                knight_moves_count += 1
+                                logging.debug(f"Knight moved from {player_clicks[0]} to {player_clicks[1]}")
+
+                            game_state.move_piece((player_clicks[0][0], player_clicks[0][1]),
+                                                  (player_clicks[1][0], player_clicks[1][1]), False)
+                            square_selected = ()
+                            player_clicks = []
+                            valid_moves = []
+
+                            if human_player == 'w':
+                                ai_move = ai.minimax_white(game_state, 3, -100000, 100000, True, Player.PLAYER_2)
+                                game_state.move_piece(ai_move[0], ai_move[1], True)
+                            elif human_player == 'b':
+                                ai_move = ai.minimax_black(game_state, 3, -100000, 100000, True, Player.PLAYER_1)
+                                game_state.move_piece(ai_move[0], ai_move[1], True)
+                    else:
+                        valid_moves = game_state.get_valid_moves((row, col)) or []
+            elif e.type == py.KEYDOWN:
+                if e.key == py.K_r:
+                    game_over = False
+                    game_state = chess_engine.game_state()
+                    square_selected = ()
+                    player_clicks = []
+                    valid_moves = []
+                elif e.key == py.K_u:
+                    game_state.undo_move()
+
+        draw_game_state(screen, game_state, valid_moves, square_selected)
+
+        endgame = game_state.checkmate_stalemate_checker()
+        if endgame == 0:
+            game_over = True
+            draw_text(screen, "Black wins.")
+        elif endgame == 1:
+            game_over = True
+            draw_text(screen, "White wins.")
+        elif endgame == 2:
+            game_over = True
+            draw_text(screen, "Stalemate.")
+
+        if game_over:
+            winner = "Black" if endgame == 0 else "White" if endgame == 1 else "Stalemate"
+            logging.info(f"Winner: {winner}")
+            logging.info(f"Total knight moves: {knight_moves_count}")
+            logging.info(f"Total CHECKs in game: {check_count}")
+
+        clock.tick(MAX_FPS)
+        py.display.flip()
+
+
+if __name__ == "__main__":
+    main()
