+++ check-the-chess-EladBibi-main/chess_log.log
@@ -1,0 +1,108 @@
+2025-05-19 23:25:30,874 - DEBUG - Knight moved from (0, 1) to (2, 2)
+2025-05-19 23:25:30,874 - INFO - Move 1: n from (0, 1) to (2, 2)
+2025-05-19 23:25:31,353 - INFO - Move 2: p from (6, 0) to (5, 0)
+2025-05-19 23:25:31,918 - INFO - Move 3: r from (0, 0) to (0, 1)
+2025-05-19 23:25:32,342 - INFO - Move 4: p from (5, 0) to (4, 0)
+2025-05-19 23:25:33,148 - INFO - Move 5: r from (0, 1) to (0, 0)
+2025-05-19 23:25:33,650 - INFO - Move 6: p from (4, 0) to (3, 0)
+2025-05-19 23:25:34,478 - INFO - Move 7: r from (0, 0) to (0, 1)
+2025-05-19 23:25:35,042 - INFO - Move 8: p from (3, 0) to (2, 0)
+2025-05-19 23:25:35,744 - INFO - Move 9: p from (1, 1) to (2, 0)
+2025-05-19 23:25:36,338 - INFO - Move 10: p from (6, 1) to (5, 1)
+2025-05-19 23:25:37,465 - INFO - Move 11: r from (0, 1) to (2, 1)
+2025-05-19 23:25:38,094 - INFO - Move 12: p from (5, 1) to (4, 1)
+2025-05-19 23:25:39,289 - INFO - Move 13: b from (0, 2) to (1, 1)
+2025-05-19 23:25:39,923 - INFO - Move 14: p from (4, 1) to (3, 1)
+2025-05-19 23:25:41,496 - INFO - Move 15: p from (2, 0) to (3, 1)
+2025-05-19 23:25:42,097 - INFO - Move 16: p from (6, 2) to (5, 2)
+2025-05-19 23:25:43,938 - INFO - Move 17: k from (0, 3) to (0, 2)
+2025-05-19 23:25:44,518 - INFO - Move 18: p from (5, 2) to (4, 2)
+2025-05-19 23:25:46,420 - INFO - Move 19: p from (3, 1) to (4, 2)
+2025-05-19 23:25:47,028 - INFO - Move 20: p from (6, 3) to (5, 3)
+2025-05-19 23:26:01,009 - INFO - Move 21: p from (4, 2) to (5, 3)
+2025-05-19 23:26:02,321 - INFO - Move 22: p from (6, 4) to (5, 4)
+2025-05-19 23:26:12,737 - INFO - Move 23: k from (0, 2) to (0, 1)
+2025-05-19 23:26:13,785 - INFO - Move 24: p from (5, 4) to (4, 4)
+2025-05-19 23:26:24,667 - DEBUG - Knight moved from (2, 2) to (3, 4)
+2025-05-19 23:26:24,667 - INFO - Move 25: n from (2, 2) to (3, 4)
+2025-05-19 23:26:26,085 - INFO - Move 26: p from (6, 5) to (5, 5)
+2025-05-19 23:26:38,580 - INFO - Move 27: q from (0, 4) to (0, 3)
+2025-05-19 23:26:40,069 - INFO - Move 28: p from (5, 5) to (4, 5)
+2025-05-19 23:30:27,791 - DEBUG - Knight moved from (0, 1) to (2, 2)
+2025-05-19 23:30:27,792 - INFO - Move 1: n from (0, 1) to (2, 2)
+2025-05-19 23:30:28,217 - INFO - Move 2: p from (6, 0) to (5, 0)
+2025-05-19 23:30:28,744 - INFO - Move 3: r from (0, 0) to (0, 1)
+2025-05-19 23:30:29,131 - INFO - Move 4: p from (5, 0) to (4, 0)
+2025-05-19 23:30:29,892 - INFO - Move 5: r from (0, 1) to (0, 0)
+2025-05-19 23:30:30,348 - INFO - Move 6: p from (4, 0) to (3, 0)
+2025-05-19 23:30:31,092 - INFO - Move 7: r from (0, 0) to (0, 1)
+2025-05-19 23:30:31,624 - INFO - Move 8: p from (3, 0) to (2, 0)
+2025-05-19 23:30:32,273 - INFO - Move 9: p from (1, 1) to (2, 0)
+2025-05-19 23:30:32,807 - INFO - Move 10: p from (6, 1) to (5, 1)
+2025-05-19 23:30:33,833 - INFO - Move 11: r from (0, 1) to (2, 1)
+2025-05-19 23:30:34,414 - INFO - Move 12: p from (5, 1) to (4, 1)
+2025-05-19 23:30:35,493 - INFO - Move 13: b from (0, 2) to (1, 1)
+2025-05-19 23:30:36,078 - INFO - Move 14: p from (4, 1) to (3, 1)
+2025-05-19 23:30:37,499 - INFO - Move 15: p from (2, 0) to (3, 1)
+2025-05-19 23:30:38,108 - INFO - Move 16: p from (6, 2) to (5, 2)
+2025-05-19 23:30:39,869 - INFO - Move 17: k from (0, 3) to (0, 2)
+2025-05-19 23:30:40,413 - INFO - Move 18: p from (5, 2) to (4, 2)
+2025-05-19 23:30:41,968 - INFO - Move 19: p from (3, 1) to (4, 2)
+2025-05-19 23:30:42,544 - INFO - Move 20: p from (6, 3) to (5, 3)
+2025-05-19 23:30:55,825 - INFO - Move 21: p from (4, 2) to (5, 3)
+2025-05-19 23:30:56,927 - INFO - Move 22: p from (6, 4) to (5, 4)
+2025-05-19 23:31:11,901 - INFO - Move 23: k from (0, 2) to (0, 1)
+2025-05-19 23:31:13,623 - INFO - Move 24: p from (5, 4) to (4, 4)
+2025-05-19 23:31:33,139 - DEBUG - Knight moved from (2, 2) to (3, 4)
+2025-05-19 23:31:33,139 - INFO - Move 25: n from (2, 2) to (3, 4)
+2025-05-19 23:31:35,686 - INFO - Move 26: p from (6, 5) to (5, 5)
+2025-05-19 23:31:49,624 - INFO - Move 27: q from (0, 4) to (0, 3)
+2025-05-19 23:31:51,006 - INFO - Move 28: p from (5, 5) to (4, 5)
+2025-05-19 23:31:55,721 - INFO - Move 29: b from (1, 1) to (4, 4)
+2025-05-19 23:31:57,339 - INFO - Move 30: p from (4, 5) to (3, 5)
+2025-05-19 23:32:25,795 - INFO - Move 31: k from (0, 1) to (1, 1)
+2025-05-19 23:32:29,349 - INFO - Move 32: p from (3, 5) to (2, 5)
+2025-05-19 23:32:57,541 - INFO - Move 33: p from (1, 4) to (2, 5)
+2025-05-19 23:32:59,267 - INFO - Move 34: p from (6, 6) to (5, 6)
+2025-05-19 23:42:40,994 - DEBUG - Knight moved from (0, 1) to (2, 2)
+2025-05-19 23:42:40,994 - INFO - Move 1: n from (0, 1) to (2, 2)
+2025-05-19 23:42:41,436 - INFO - Move 2: p from (6, 0) to (5, 0)
+2025-05-19 23:42:41,909 - INFO - Move 3: r from (0, 0) to (0, 1)
+2025-05-19 23:42:42,297 - INFO - Move 4: p from (5, 0) to (4, 0)
+2025-05-19 23:42:43,229 - INFO - Move 5: r from (0, 1) to (0, 0)
+2025-05-19 23:42:43,695 - INFO - Move 6: p from (4, 0) to (3, 0)
+2025-05-19 23:42:44,429 - INFO - Move 7: r from (0, 0) to (0, 1)
+2025-05-19 23:42:44,960 - INFO - Move 8: p from (3, 0) to (2, 0)
+2025-05-19 23:42:45,584 - INFO - Move 9: p from (1, 1) to (2, 0)
+2025-05-19 23:42:46,144 - INFO - Move 10: p from (6, 1) to (5, 1)
+2025-05-19 23:42:47,143 - INFO - Move 11: r from (0, 1) to (2, 1)
+2025-05-19 23:42:47,728 - INFO - Move 12: p from (5, 1) to (4, 1)
+2025-05-19 23:42:48,806 - INFO - Move 13: b from (0, 2) to (1, 1)
+2025-05-19 23:42:49,374 - INFO - Move 14: p from (4, 1) to (3, 1)
+2025-05-19 23:42:50,792 - INFO - Move 15: p from (2, 0) to (3, 1)
+2025-05-19 23:42:51,343 - INFO - Move 16: p from (6, 2) to (5, 2)
+2025-05-19 23:42:53,027 - INFO - Move 17: k from (0, 3) to (0, 2)
+2025-05-19 23:42:53,585 - INFO - Move 18: p from (5, 2) to (4, 2)
+2025-05-19 23:42:55,143 - INFO - Move 19: p from (3, 1) to (4, 2)
+2025-05-19 23:42:56,000 - INFO - Move 20: p from (6, 3) to (5, 3)
+2025-05-19 23:43:10,210 - INFO - Move 21: p from (4, 2) to (5, 3)
+2025-05-19 23:43:11,259 - INFO - Move 22: p from (6, 4) to (5, 4)
+2025-05-19 23:43:21,541 - INFO - Move 23: k from (0, 2) to (0, 1)
+2025-05-19 23:43:22,514 - INFO - Move 24: p from (5, 4) to (4, 4)
+2025-05-19 23:43:33,345 - DEBUG - Knight moved from (2, 2) to (3, 4)
+2025-05-19 23:43:33,346 - INFO - Move 25: n from (2, 2) to (3, 4)
+2025-05-19 23:43:34,709 - INFO - Move 26: p from (6, 5) to (5, 5)
+2025-05-19 23:43:46,209 - INFO - Move 27: q from (0, 4) to (0, 3)
+2025-05-19 23:43:47,664 - INFO - Move 28: p from (5, 5) to (4, 5)
+2025-05-19 23:43:52,455 - INFO - Move 29: b from (1, 1) to (4, 4)
+2025-05-19 23:43:54,205 - INFO - Move 30: p from (4, 5) to (3, 5)
+2025-05-19 23:44:29,747 - INFO - Move 31: k from (0, 1) to (1, 1)
+2025-05-19 23:44:33,750 - INFO - Move 32: p from (3, 5) to (2, 5)
+2025-05-19 23:45:24,470 - INFO - Move 33: p from (1, 4) to (2, 5)
+2025-05-19 23:45:27,930 - INFO - Move 34: p from (6, 6) to (5, 6)
+2025-05-19 23:46:38,339 - INFO - Move 35: b from (4, 4) to (7, 7)
+2025-05-19 23:46:41,233 - INFO - Move 36: p from (5, 6) to (4, 6)
+2025-05-19 23:47:11,481 - INFO - Move 37: b from (0, 5) to (3, 2)
+2025-05-19 23:47:14,811 - INFO - Move 38: p from (4, 6) to (3, 6)
+2025-05-19 23:47:24,696 - INFO - Move 39: b from (3, 2) to (7, 6)
+2025-05-19 23:47:26,173 - INFO - Move 40: p from (3, 6) to (2, 6)
