+++ check-the-chess-EladBibi-main/enums.py
@@ -1,0 +1,6 @@
+class Player:
+    PLAYER_1 = 'white'
+    PLAYER_2 = 'black'
+    EMPTY = -9
+    PIECES = ['white_r', 'white_n', 'white_b', 'white_q', 'white_k', 'white_p',
+              'black_r', 'black_n', 'black_b', 'black_q', 'black_k', 'black_p']
